Implemented Player Model and "Animations"
Player model is a blender model, so to rotate on the Y axis, you gotta
rotate on its Z axis.

I made the PlayerAnimation script with public functions that will do an
animation for a preset amount of time.

If you want to have a set face, you have to do that yourself using an
array of gameobjects which are the faces.