SHMUP Win State
There is a win state! You can win. And then go back to the HubScene.

GameManager carries over all the data. Which is....the score and that is it

But you can go from the Hub World to the SCHMUP and keep raising that score

so yeah WE ARE GOING PLACES